lint fixes and refactoring
